requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.13.3
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
httpx==0.27.2
requests==2.32.3
python-dotenv==1.0.1
numpy==2.1.2
librosa==0.10.2.post1
soundfile==0.12.1
Pillow==10.4.0
opencv-python-headless==4.10.0.84
scikit-learn==1.5.2
tqdm==4.66.5
